GPIO_OUTPUT_INACTIVE makes regulator_common_init() use delay
Commit https://github.com/zephyrproject-rtos/zephyr/commit/7fa815c92967f26a7c9af33776677015013e18d0 introduces a variable, that is basically always true. Because of that regulator_common_init() never chooses the previously chosen `else if` path and doesn't wait the defined startup-delay-us time. Hence, connected devices don't have enough time to power up.

This fix is NOT a real fix. It works for me, but as soon as the enable-pin
was already handled by a bootloader, those pins will be configured a
inactive outputs. If the bootloader already actived the pin, it will be
reset before it's actived once more.
This is a structural issue. regulator-boot-on/off is checked after
configuring the pin.